src/test/kotlin/com/bhkpo/servlet/domain/member/MemberRepositoryTest.kt
package com.bhkpo.servlet.domain.member

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.AfterEach
import org.junit.jupiter.api.Test

class MemberRepositoryTest {
    private val memberRepository = MemberRepository.getInstance()

    @AfterEach
    fun afterEach() {
        memberRepository.clearStore()
    }

    @Test
    fun save() {
        // given
        val member = Member("hello", 20)

        // when
        val savedMember = memberRepository.save(member)

        // then
        val findMember: Member? = memberRepository.findById(savedMember.id)
        assertThat(findMember).isEqualTo(savedMember)
    }

    @Test
    fun findAll() {
        // given
        val member1 = Member("member1", 20)
        val member2 = Member("member2", 30)

        memberRepository.save(member1)
        memberRepository.save(member2)

        // when
        val result: List<Member> = memberRepository.findAll()

        // then
        assertThat(result.size).isEqualTo(2)
        assertThat(result).contains(member1, member2)
    }
}
